requirements.txt
asgiref==3.2.10
attrs==20.1.0
Django==3.1
django-debug-toolbar==2.2
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
freezegun==0.3.15
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytest-freezegun==0.4.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
